move llm, embedder, vectordb items to components folder
